+++ app/src/main/java/com/example/todolist/view/compose_components/NotesList.kt
@@ -1,4 +1,4 @@
-package com.example.todolist.view.composecomponents
+package com.example.todolist.view.compose_components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
