+++ app/src/main/java/com/example/todolist/view/screens/MainScreen.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.example.todolist.R
 import com.example.todolist.data.local.Note
-import com.example.todolist.view.composecomponents.NotesList
-import com.example.todolist.view.composecomponents.TopBar
+import com.example.todolist.view.compose_components.NotesList
+import com.example.todolist.view.compose_components.TopBar
 import com.example.todolist.viewmodel.NotesViewModel
 
 @Composable
